+++ Domain/Entities/Recibo.cs
@@ -4,7 +4,7 @@
     {
         public Guid IdRecibo { get; set; }
         public Guid IdDescuento { get; set; }
-        public double precioTotal { get; set; }
+        public decimal precioTotal { get; set; }
 
         public ICollection<Pedido> Pedidos { get; set; }
         public Descuento descuento { get; set; }
